@@ -21,87 +22,86 @@
     "prisma:generate:watch": "prisma generate --watch",
     "storybook": "nuxt storybook",
     "storybook:build": "nuxt storybook build",
+    "storybook@17": "cross-env NODE_OPTIONS=--openssl-legacy-provider nuxt storybook",
+    "storybook:build@17": "cross-env NODE_OPTIONS=--openssl-legacy-provider nuxt storybook build",
     "storybook:publish": "chromatic --exit-zero-on-changes --build-script-name storybook:build",
-    "graphql:generate": "cross-env TS_NODE_PROJECT=./tsconfig.cjs.json graphql-codegen",
-    "graphql:generate:watch": "cross-env TS_NODE_PROJECT=./tsconfig.cjs.json graphql-codegen --watch",
+    "graphql:generate": "graphql-codegen",
+    "graphql:generate:watch": "graphql-codegen --watch",
     "graphql:validate": "graphql-inspector validate --apollo '{pages,components}/**/*.vue' 'api/**/*.graphql'",
     "docker:redis": "docker run -d --rm --name JabRefRedis -p 6380:6379 redis --requirepass jabref",
     "lint": "yarn lint:js && yarn lint:graphql",
     "lint:js": "eslint --ext .ts,.js,.vue --ignore-path .gitignore .",
-    "lint:graphql": "cross-env TS_NODE_PROJECT=./tsconfig.cjs.json eslint --ext .graphql,.vue --ignore-path .gitignore .",
+    "lint:graphql": "eslint --ext .graphql,.vue --ignore-path .gitignore .",
     "typecheck": "vue-tsc --noEmit",
-    "postinstall": "yarn generate",
+    "postinstall": "patch-package && nuxi prepare && yarn generate",
     "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --runInBand",
     "test:integration": "cross-env NODE_OPTIONS=--experimental-vm-modules jest test.ts --runInBand",
     "test:unit": "cross-env NODE_OPTIONS=--experimental-vm-modules jest spec.ts",
     "validate": "yarn graphql:validate"
   },
   "dependencies": {
-    "@apollo/client": "^3.4.17",
-    "@babel/plugin-proposal-decorators": "^7.16.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.16.0",
+    "@apollo/client": "^3.5.6",
+    "@babel/plugin-proposal-decorators": "^7.16.7",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.7",
+    "@babel/plugin-proposal-optional-chaining": "^7.16.7",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
-    "@babel/plugin-transform-modules-commonjs": "^7.16.0",
-    "@babel/plugin-transform-typescript": "^7.16.1",
+    "@babel/plugin-transform-modules-commonjs": "^7.16.7",
+    "@babel/plugin-transform-typescript": "^7.16.7",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/vue-fontawesome": "^2.0.6",
     "@graphql-tools/graphql-file-loader": "^7.3.3",
-    "@graphql-tools/load": "^7.4.1",
+    "@graphql-tools/load": "^7.5.1",
     "@graphql-tools/merge": "^8.2.1",
     "@graphql-tools/schema": "^8.3.0",
-    "@he-tree/vue2": "^1.3.5",
-    "@nuxtjs/composition-api": "^0.30.0",
-    "@pinia/nuxt": "^0.1.0",
-    "@prisma/client": "^3.4.2",
-    "@vue/apollo-composable": "4.0.0-alpha.15",
-    "@vue/apollo-util": "^4.0.0-alpha.15",
-    "@vueuse/core": "^6.8.0",
-    "@yaireo/tagify": "^4.8.1",
-    "apollo-server-express": "^3.5.0",
-    "autoprefixer": "^10.4.0",
-    "babel-plugin-transform-import-meta": "^2.1.0",
-    "babel-plugin-transform-typescript": "^7.0.0-alpha.19",
-    "babel-plugin-transform-typescript-metadata": "^0.3.2",
+    "@he-tree/vue2": "^1.3.9",
+    "@pinia/nuxt": "^0.1.8",
+    "@prisma/client": "^3.7.0",
+    "@vue/apollo-composable": "4.0.0-alpha.16",
+    "@vue/apollo-util": "^4.0.0-alpha.16",
+    "@vueuse/core": "^7.5.3",
+    "@yaireo/tagify": "^4.9.4",
+    "apollo-server-express": "^3.6.1",
+    "autoprefixer": "^10.4.2",
     "connect-redis": "^6.0.0",
     "cross-fetch": "^3.1.4",
-    "eslint-plugin-jest": "^25.2.4",
+    "eslint-plugin-jest": "^25.3.4",
     "eslint-plugin-prettier-vue": "^3.1.0",
-    "express": "^4.17.1",
+    "express": "^4.17.2",
     "express-session": "^1.17.2",
-    "graphql": "^15.7.2",
+    "graphql": "^15.8.0",
     "graphql-passport": "^0.6.4",
-    "graphql-scalars": "^1.12.0",
+    "graphql-scalars": "^1.14.1",
     "lodash": "^4.17.21",
-    "nodemailer": "^6.7.0",
-    "nuxt": "^2.15.8",
-    "nuxt-route-meta": "^2.3.2",
-    "passport": "^0.5.0",
-    "pinia": "^2.0.3",
+    "nodemailer": "^6.7.2",
+    "nuxt-edge": "latest",
+    "nuxt-route-meta": "^2.3.4",
+    "passport": "^0.5.2",
+    "pinia": "^2.0.9",
     "portal-vue": "^2.1.7",
-    "redis": "^3.1.2",
+    "redis": "^4.0.1",
     "reflect-metadata": "^0.1.13",
     "ts-loader": "^8.3.0",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.4.0",
     "tsyringe": "^4.6.0",
-    "typescript": "^4.4.4",
-    "vue-apollo": "^3.0.8",
+    "typescript": "^4.5.4",
+    "vue-apollo": "^3.1.0",
     "vue-tailwind": "^2.5.0",
     "vue-virtual-scroll-list": "^2.3.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@graphql-codegen/cli": "^2.2.2",
-    "@graphql-codegen/fragment-matcher": "^3.2.0",
-    "@graphql-codegen/gql-tag-operations-preset": "^1.2.1",
-    "@graphql-codegen/typed-document-node": "^2.2.0",
+    "@babel/core": "^7.16.7",
+    "@graphql-codegen/cli": "^2.3.1",
+    "@graphql-codegen/fragment-matcher": "^3.2.1",
+    "@graphql-codegen/gql-tag-operations-preset": "^1.2.2",
+    "@graphql-codegen/typed-document-node": "^2.2.2",
     "@graphql-codegen/typescript": "^2.2.4",
-    "@graphql-codegen/typescript-operations": "^2.2.0",
-    "@graphql-codegen/typescript-resolvers": "^2.4.0",
-    "@graphql-eslint/eslint-plugin": "^2.4.1",
+    "@graphql-codegen/typescript-operations": "^2.2.2",
+    "@graphql-codegen/typescript-resolvers": "^2.4.3",
+    "@graphql-eslint/eslint-plugin": "^3.6.0",
     "@graphql-inspector/cli": "^3.0.2",
-    "@graphql-typed-document-node/core": "^3.1.0",
+    "@graphql-typed-document-node/core": "^3.1.1",
+    "@nuxt/bridge": "npm:@nuxt/bridge-edge",
     "@nuxt/postcss8": "^1.1.3",
     "@nuxt/types": "^2.15.8",
     "@nuxt/typescript-build": "^2.0.6",
