@@ -78,7 +78,9 @@ export class Mutation {
         problems: [
           {
             path: 'Email or Password',
-            message: (typeof info === 'string' ? info : info?.message) || 'Unknown error while logging in.',
+            message:
+              (typeof info === 'string' ? info : info?.message) ||
+              'Unknown error while logging in.',
           },
         ],
       }
