@@ -15,8 +15,8 @@
   </span>
 </template>
 
-<script>
-import { defineComponent } from '@nuxtjs/composition-api'
+<script lang="ts">
+import { defineComponent } from '@vue/composition-api'
 export default defineComponent({
   props: ['value'],
   methods: {
