@@ -1,16 +1,7 @@
 <template>
   <nav class="sticky top-0 z-50 flex-none">
     <div
-      class="
-        w-full
-        flex flex-row flex-wrap
-        items-center
-        justify-between
-        bg-white
-        px-6
-        border-b border-gray-300
-        h-20
-      "
+      class="w-full flex flex-row flex-wrap items-center justify-between bg-white px-6 border-b border-gray-300 h-20"
     >
       <div class="flex">
         <!-- Hamburger icon for small screens -->
@@ -26,16 +17,7 @@
             v-model="searchQuery"
             type="search"
             placeholder="Search..."
-            class="
-              w-full
-              xl:w-96
-              px-5
-              border-none
-              shadow-none
-              pl-10
-              font-semibold
-              text-lg
-            "
+            class="w-full xl:w-96 px-5 border-none shadow-none pl-10 font-semibold text-lg"
           />
           <button type="submit" class="absolute left-0 top-0 bottom-0 ml-2">
             <FontAwesomeIcon icon="search" class="text-gray-400" />
@@ -96,61 +78,21 @@
 
             <div slot-scope="{ blurHandler }" class="w-36">
               <button
-                class="
-                  block
-                  w-full
-                  px-4
-                  py-2
-                  text-sm
-                  leading-5
-                  text-gray-700
-                  transition
-                  duration-150
-                  ease-in-out
-                  hover:bg-gray-100
-                  focus:outline-none focus:bg-gray-100
-                "
+                class="block w-full px-4 py-2 text-sm leading-5 text-gray-700 transition duration-150 ease-in-out hover:bg-gray-100 focus:outline-none focus:bg-gray-100"
                 role="menuitem"
                 @blur="blurHandler"
               >
                 Your Profile
               </button>
               <button
-                class="
-                  block
-                  w-full
-                  px-4
-                  py-2
-                  text-sm
-                  leading-5
-                  text-gray-700
-                  transition
-                  duration-150
-                  ease-in-out
-                  hover:bg-gray-100
-                  focus:outline-none focus:bg-gray-100
-                "
+                class="block w-full px-4 py-2 text-sm leading-5 text-gray-700 transition duration-150 ease-in-out hover:bg-gray-100 focus:outline-none focus:bg-gray-100"
                 role="menuitem"
                 @blur="blurHandler"
               >
                 Settings
               </button>
               <button
-                class="
-                  block
-                  w-full
-                  px-4
-                  py-2
-                  text-sm
-                  leading-5
-                  text-gray-700
-                  transition
-                  duration-150
-                  ease-in-out
-                  border-t
-                  hover:bg-gray-100
-                  focus:outline-none focus:bg-gray-100
-                "
+                class="block w-full px-4 py-2 text-sm leading-5 text-gray-700 transition duration-150 ease-in-out border-t hover:bg-gray-100 focus:outline-none focus:bg-gray-100"
                 @click="logout"
               >
                 Logout
@@ -165,7 +107,8 @@
 
 <script lang="ts">
 import { useApolloClient, useMutation } from '@vue/apollo-composable'
-import { defineComponent, useRouter, ref, watch } from '@nuxtjs/composition-api'
+import { defineComponent, ref, watch } from '@vue/composition-api'
+import { useRouter } from '#app'
 import { gql } from '~/apollo'
 import { useUiStore } from '~/store'
 import { cacheCurrentUser } from '~/apollo/cache'
