@@ -16,7 +16,7 @@
   </section>
 </template>
 
-<script>
+<script lang="ts">
 export default {
   layout: 'bare',
   props: {
