@@ -1,49 +1,48 @@
 import Vue from 'vue'
 import VueApollo from 'vue-apollo'
 import { ApolloClient, HttpLink } from '@apollo/client/core'
-import { Plugin } from '@nuxt/types'
 import fetch from 'cross-fetch'
-import { DefaultApolloClient } from '@vue/apollo-composable'
 import { onError } from '@apollo/client/link/error'
 import { logErrorMessages } from '@vue/apollo-util'
-import { provide, onGlobalSetup } from '@nuxtjs/composition-api'
+import { provideApolloClient } from '@vue/apollo-composable'
 import { cache } from '../apollo/cache'
-import { config, Environment } from '~/config'
+import { defineNuxtPlugin, useRuntimeConfig } from '#app'
+import { Environment } from '~/config'
 
 Vue.use(VueApollo)
 
-let httpLink
-if (config.environment === Environment.Production) {
-  httpLink = new HttpLink({ uri: '/api', fetch })
-} else {
-  httpLink = new HttpLink({ uri: 'http://localhost:3000/api', fetch })
-}
-
-// Print errors
-const errorLink = onError((error) => {
-  if (config.environment !== Environment.Production) {
-    logErrorMessages(error)
+export default defineNuxtPlugin((nuxtApp) => {
+  if (!nuxtApp) {
+    // For some strange reason, nuxtApp is not defined for storybook, so don't do anything in this case
+    return
   }
-})
 
-// Create the apollo client
-const apolloClient = new ApolloClient({
-  cache,
-  link: errorLink.concat(httpLink),
-  // Send cookies along with every request (needed for authentication)
-  credentials: 'include',
-})
+  const config = useRuntimeConfig()
+  let httpLink
+  if (config.environment === Environment.Production) {
+    httpLink = new HttpLink({ uri: '/api', fetch })
+  } else {
+    httpLink = new HttpLink({ uri: 'http://localhost:3000/api', fetch })
+  }
 
-const apolloProvider = new VueApollo({
-  defaultClient: apolloClient,
-})
-
-const apolloPlugin: Plugin = function ({ app }, _inject) {
-  app.apolloProvider = apolloProvider
-
-  onGlobalSetup(() => {
-    provide(DefaultApolloClient, apolloProvider.defaultClient)
+  // Print errors
+  const errorLink = onError((error) => {
+    if (config.environment !== Environment.Production) {
+      logErrorMessages(error)
+    }
   })
-}
 
-export default apolloPlugin
+  // Create the apollo client
+  const apolloClient = new ApolloClient({
+    cache,
+    link: errorLink.concat(httpLink),
+    // Send cookies along with every request (needed for authentication)
+    credentials: 'include',
+  })
+
+  const apolloProvider = new VueApollo({
+    defaultClient: apolloClient,
+  })
+
+  provideApolloClient(apolloProvider?.defaultClient)
+})
