@@ -1,13 +1,16 @@
-import { RedisClient } from 'redis'
+import { RedisClientType } from 'redis'
 import prisma from '@prisma/client'
 import { container, instanceCachingFactory } from 'tsyringe'
+import dotenv from 'dotenv'
 import { createRedisClient } from '~/api/utils/services.factory'
 
 // Register services for all tests
-container.register(RedisClient, {
-  useFactory: instanceCachingFactory(() => createRedisClient()),
+container.register('RedisClient', {
+  useValue: await createRedisClient(),
 })
-afterAll(() => container.resolve(RedisClient).quit())
+afterAll(async () => {
+  await container.resolve<RedisClientType>('RedisClient').quit()
+})
 
 // Setup services for integration tests
 // @ts-ignore: Jest doesn't allow an easy way to add typescript info
@@ -16,3 +19,6 @@ if (global.isIntegrationTest) {
     useFactory: instanceCachingFactory(() => new prisma.PrismaClient()),
   })
 }
+
+// Load environment variables from .env file
+dotenv.config()
