@@ -12,7 +12,7 @@ declare module 'vue-virtual-scroll-list' {
 // Allows to import vue files
 // Taken from https://github.com/vuejs/vue/issues/5298#issuecomment-761577986
 declare module '*.vue' {
-  import { defineComponent } from '@nuxtjs/composition-api'
+  import { defineComponent } from '@vue/composition-api'
   const component: ReturnType<typeof defineComponent>
   export default component
 }
